requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
chromadb==0.4.18
numpy<2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
python-multipart==0.0.6
pytz==2023.3

